requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
customtkinter>=5.2.0
ultralytics>=8.0.0
requests>=2.31.0
tqdm>=4.66.0
roboflow>=1.1.0
kaggle>=1.6.0
pytest
pytest-cov
